ml-vision/requirements.txt
google-cloud-vision
google-cloud-storage
#dlib
#imutils
#numpy
#opencv-python

# sudo apt-get install libx11-dev libgtk-3-dev
# sudo apt-get install libopenblas-dev liblapack-dev 

# https://www.pyimagesearch.com/2018/01/22/install-dlib-easy-complete-guide/
# Installing dlib in target environment